pdfImporter/requirements.txt
easyocr>=1.6.0
PyPDF2>=3.0.0
pdf2image
PyMuPDF>=1.19.0
numpy>=1.22.0
opencv-python>=4.5.5
Pillow>=9.0.0